Fix warnings from clang

+++ src/solution.h
@@ -6,13 +6,13 @@ class Solution {
     ElectronDensityDistribution *model;
     std::vector<Grid> grids;
     std::vector<std::vector<Ray>> data;
-    float latitudeLeft, longitudeLeft, timeLeft;
-    float latitudeRight, longitudeRight, timeRight;
+    double latitudeLeft, longitudeLeft, timeLeft;
+    double latitudeRight, longitudeRight, timeRight;
 public:
     Solution();
-    void setLimits(float latitudeLeft, float latitudeRight, float longitudeLeft, float longitudeRight, float timeLeft, float timeRight);
+    void setLimits(double latitudeLeft, double latitudeRight, double longitudeLeft, double longitudeRight, double timeLeft, double timeRight);
     void setModel(ElectronDensityDistribution &model);
-    void addGrid(int spaceIntervals, int timeIntervals);
+    void addGrid(unsigned spaceIntervals, unsigned timeIntervals);
     void addData(std::vector<std::vector<Ray>> data);
     void find();
     void print();
